Equal Frequency
Equal Frequency seems to work. I mean, trying the all possible cutpoints
and computing the sum of the sqaure of the distances from the ideal
length crashed to program doing 5 cutpoints on 200 data points..
soooo... instead I use the metric of trying to make the first interval
most closely approach the ideal length recursively. This method at least
got the same result as I did on HW2 for the equal frequency
discretization.